requirements.txt
pySerial

